HealthBar Update: QOL change
- added hp value to hp bar for easier indication of how much hp an enemy has

+++ Ecs657/Assets/Scripts/HealthBar.cs
@@ -9,7 +9,8 @@ public class HealthBar : MonoBehaviour
     [SerializeField] Gradient gradient;
     // Start is called before the first frame update
 
-    public void setMaxHealth(int maxHealth)
+    //increase max hp and set HP to that amount
+    public void setMaxHealth(float maxHealth)
 	{
         slider.maxValue = maxHealth;
         slider.value = maxHealth;
